chore: space 관련 admin API를 admin 모듈로 이전
