+++ layer-api/src/main/java/org/layer/domain/admin/controller/AdminController.java
@@ -3,8 +3,6 @@ package org.layer.domain.admin.controller;
 import lombok.RequiredArgsConstructor;
 import org.layer.domain.admin.controller.dto.AdminRetrospectCountGetResponse;
 import org.layer.domain.admin.controller.dto.AdminRetrospectsGetResponse;
-import org.layer.domain.admin.controller.dto.AdminSpaceCountGetResponse;
-import org.layer.domain.admin.controller.dto.AdminSpacesGetResponse;
 import org.layer.domain.admin.service.AdminService;
 import org.layer.domain.retrospect.dto.AdminRetrospectCountGroupBySpaceGetResponse;
 import org.springframework.http.ResponseEntity;
@@ -19,15 +17,6 @@ import java.util.List;
 public class AdminController implements AdminApi {
 	private final AdminService adminService;
 
-	@Override
-	@GetMapping("/space")
-	public ResponseEntity<AdminSpacesGetResponse> getSpaceData(
-		@RequestParam("startDate") LocalDateTime startDate,
-		@RequestParam("endDate") LocalDateTime endDate,
-		@RequestParam("password") String password) {
-
-		return ResponseEntity.ok(adminService.getSpaceData(startDate, endDate, password));
-	}
 
 	@Override
 	@GetMapping("/retrospect")
@@ -39,14 +28,7 @@ public class AdminController implements AdminApi {
 		return ResponseEntity.ok(adminService.getRetrospectData(startDate, endDate, password));
 	}
 
-	@Override
-	@GetMapping("/space/count/user-only")
-	public ResponseEntity<AdminSpaceCountGetResponse> getSpaceCount(
-			@RequestParam("startDate") LocalDateTime startDate,
-			@RequestParam("endDate") LocalDateTime endDate,
-			@RequestParam("password") String password) {
-		return ResponseEntity.ok(adminService.getSpaceCount(startDate, endDate, password));
-	}
+
 
 	@Override
 	@GetMapping("/retrospect/count/user-only")
